always print the final AST
The previous implementation would always print the final AST if verbose
mode was set. If an output file was specified, it would also print the
AST into the file.

However, in the case where verbose mode was disabled and no output file
was specified, the AST would not be printed. The current implementation
fixes this.